+++ src/components/Detail/MusicAbout/MusicCommentList/CommentTextArea/CommentTextArea.scss
@@ -1,0 +1,3 @@
+.music-comment-list__input {
+  resize: none;
+}
