+++ src/components/Detail/MusicAbout/MusicCommentList/CommentTextArea/index.ts
@@ -1,0 +1,3 @@
+import CommentTextArea from "./CommentTextArea";
+
+export default CommentTextArea;
